refactor: renamed widget -> view
