+++ src/tagstudio/qt/views/widgets/search_bar_view.py
